Check whether converted object is null; make ContentPropertyInfoConverterAdapter private
